+++ packages/flame/lib/src/components/input/hud_button_component.dart
@@ -6,8 +6,15 @@ import '../../../extensions.dart';
 import '../../../input.dart';
 import '../../gestures/events.dart';
 
+/// The [HudButtonComponent] bundles two [PositionComponent]s, one that shows
+/// when the button is being pressed, and one that shows otherwise.
+///
+/// Note: if you are setting the [button] in [onLoad] instead of passing it in
+/// through the constructor, be sure that you add it to the component too by
+/// simply doing `add(button);` and that you also set the [button] variable to
+/// your [PositionComponent] with `button = yourPositionComponent;`.
 class HudButtonComponent extends HudMarginComponent with Tappable {
-  late final PositionComponent button;
+  late final PositionComponent? button;
   late final PositionComponent? buttonDown;
 
   /// Callback for what should happen when the button is pressed.
@@ -16,31 +23,41 @@ class HudButtonComponent extends HudMarginComponent with Tappable {
   VoidCallback? onPressed;
 
   HudButtonComponent({
-    required this.button,
+    this.button,
     this.buttonDown,
     EdgeInsets? margin,
+    this.onPressed,
     Vector2? position,
     Vector2? size,
-    Anchor anchor = Anchor.center,
-    this.onPressed,
+    Vector2? scale,
+    double? angle,
+    Anchor? anchor,
+    int? priority,
   }) : super(
           margin: margin,
           position: position,
-          size: size ?? button.size,
+          size: size ?? button?.size,
+          scale: scale,
+          angle: angle,
           anchor: anchor,
+          priority: priority,
         );
 
   @override
   Future<void> onLoad() async {
     await super.onLoad();
-    add(button);
+    if (button != null) {
+      add(button!);
+    }
   }
 
   @override
   @mustCallSuper
   bool onTapDown(TapDownInfo info) {
     if (buttonDown != null) {
-      children.remove(button);
+      if (button != null) {
+        remove(button!);
+      }
       add(buttonDown!);
     }
     onPressed?.call();
@@ -58,8 +75,10 @@ class HudButtonComponent extends HudMarginComponent with Tappable {
   @mustCallSuper
   bool onTapCancel() {
     if (buttonDown != null) {
-      children.remove(buttonDown!);
-      add(button);
+      remove(buttonDown!);
+      if (button != null) {
+        add(button!);
+      }
     }
     return false;
   }